+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 *	fix: handle case when cached key/value are obsolete after refactoring
 *	fix: handle case when @Cacheable method have no argument
+*	fix: handle generic return type for cache key
 
 ## 0.4.1 - 2022-02-14